split up lib.rs into smaller, easier to look thru parts
